+++ volttron/platform/aip.py
@@ -61,8 +61,7 @@ try:
 except ImportError:
     import json as jsonapi
 
-from .agent.utils import is_valid_identity, get_messagebus, \
-    get_platform_instance_name
+from .agent.utils import is_valid_identity, get_messagebus
 from .packages import UnpackedPackage
 from .vip.agent import Agent
 from .keystore import KeyStore
@@ -662,17 +661,15 @@ class AIPplatform(object):
         # auth only cert created is used
         msg_bus = get_messagebus()
         if msg_bus == 'rmq':
-            instance_name = get_platform_instance_name()
-            rmq_user = instance_name + '.' + agent_vip_identity
-            config_access = "{user}|{user}.pubsub.*|{user}.zmq.*".format(user=rmq_user)
+            config_access = "{identity}|{identity}.pubsub.*|{identity}.zmq.*".format(identity=agent_vip_identity)
             read_access = "volttron|{}".format(config_access)
             write_access = "volttron|{}".format(config_access)
             permissions = dict(configure=config_access, read=read_access, write=write_access)
             ssl = is_ssl_connection()
             if ssl:
                 _log.info("Created agent cert")
-                create_vagent_cert(rmq_user)
-            create_rmq_user_with_permissions(rmq_user, permissions)
+                create_vagent_cert(agent_vip_identity)
+            create_rmq_user_with_permissions(agent_vip_identity, permissions)
 
         module, _, func = module.partition(':')
         if func:
