+++ volttron/utils/rmq_setup.py
@@ -63,10 +63,7 @@ from rmq_mgmt import is_ssl_connection, get_vhost, http_put_request, \
     is_valid_mgmt_port, init_rabbitmq_setup, get_ssl_url_params, create_user, \
     set_user_permissions, set_parameter
 
-try:
-    import yaml
-except ImportError:
-    raise RuntimeError('PyYAML must be installed before running this script ')
+#disable_warnings(exceptions.SecurityWarning)
 
 _log = logging.getLogger(__name__)
 
@@ -80,7 +77,6 @@ admin_user = None  # User to prompt for if we go the docker route
 admin_password = None
 rabbitmq_server = 'rabbitmq_server-3.7.7'
 
-
 def _load_rmq_config(volttron_home=None):
     """
     Load RabbitMQ config from VOLTTRON_HOME
@@ -91,216 +87,389 @@ def _load_rmq_config(volttron_home=None):
     global config_opts, volttron_rmq_config
     if not volttron_home:
         volttron_home = get_home()
-    try:
-        volttron_rmq_config = os.path.join(volttron_home, 'rabbitmq_config.yml')
-        with open(volttron_rmq_config, 'r') as yaml_file:
-            config_opts = yaml.load(yaml_file)
-    except IOError as exc:
-        print "Error opening {}".format(volttron_rmq_config)
-        return exc
-    except yaml.YAMLError as exc:
-        return exc
+    if not os.path.exists(volttron_home):
+        os.makedirs(volttron_home)
+    volttron_rmq_config = os.path.join(volttron_home, 'rabbitmq_config.json')
+    config_opts = PersistentDict(filename=volttron_rmq_config, flag='c',
+                                 format='json')
 
 
-def _check_basic_rabbit_config():
+def _create_federation_setup():
     """
-    Check if basic rabbitmq configuration is available.
+    Creates a RabbitMQ federation of multiple VOLTTRON instances.
+    That means it creates upstream servers and sets "volttron" exchange to be "federated".
+
     :return:
     """
     global config_opts
-    error = _load_rmq_config()
-    if error:
-        return error
+    if not config_opts:
+        _load_rmq_config()
+
+    federation = config_opts['federation-upstream']
+
+    for name, address in federation.iteritems():
+
+        property = dict(vhost='volttron',
+                        component="federation-upstream",
+                        name=name,
+                        value={"uri":address})
+        set_parameter('federation-upstream', name, property,
+                      config_opts['virtual-host'])
+
+    policy_name = 'volttron-federation'
+    policy_value = {"pattern":"^volttron",
+                    "definition": {"federation-upstream-set":"all"},
+                    "priority":0,
+                    "apply-to": "exchanges"}
+    set_policy(policy_name, policy_value,
+               config_opts['virtual-host'])
+
+
+def _set_initial_rabbit_config(instance_name):
+    """
+    Build rabbitmq config for a single instance based on user input
+    :param instance_name:
+    :return:
+    """
+    global config_opts
+    _load_rmq_config()
+
     rmq_home = config_opts.get("rmq-home")
-    if not rmq_home:
-        rmq_home = os.path.join(os.path.expanduser("~"), "rabbitmq_server/rabbitmq_server-3.7.7")
-        if os.path.exists(rmq_home):
-            config_opts.setdefault("rmq-home", rmq_home)
-        else:
-            print("Missing Key in RabbitMQ config. RabbitMQ is not installed in default path: \n"
-                  "~/rabbitmq_server/rabbitmq_server-3.7.7 \n"
-                  "Set the correct RabbitMQ installation path")
-            return False
+    default_dir = os.path.join(os.path.expanduser("~"), "rabbitmq_server")
+    if rmq_home:
+        rmq_install_dir = os.path.dirname(rmq_home)
+    else:
+        rmq_install_dir = default_dir
+    valid_dir = False
+    while not valid_dir:
+        prompt = 'Rabbitmq install directory:'
+        rmq_install_dir = prompt_response(prompt, default=rmq_install_dir)
+        if rmq_install_dir == default_dir and not os.path.exists(default_dir):
+            os.mkdir(default_dir)
+        valid_dir = os.access(rmq_install_dir, os.W_OK)
+        if not valid_dir:
+            print ("Invalid install directory. Directory should exist and "
+                   "should have write access to user")
 
-    # Check if basic configuration is available in the config file, if not set default.
-    config_opts.setdefault('host',"localhost")
-    config_opts.setdefault("ssl", "true")
-    config_opts["amqp-port"] = amqp_port = 5672 # TODO - If user defined port, this needs to change
-    config_opts["mgmt-port"] = mgmt_port = 15672 # TODO - If user defined port, this needs to change
-    config_opts.setdefault('virtual-host', "volttron")
-
-    with open("{vhome}/rabbitmq_config.yml".format(vhome=get_home()), 'w') as yaml_file:
-        yaml.dump(config_opts, yaml_file, default_flow_style=False)
-
+    rmq_home = os.path.join(rmq_install_dir, rabbitmq_server)
     if os.path.exists(rmq_home) and \
             os.path.exists(os.path.join(rmq_home, 'sbin/rabbitmq-server')):
+        print("Given directory already contains {}. "
+              "Skipping rabbitmq server install".format(rabbitmq_server))
         # if existing server attempt to stop
         stop_rabbit(rmq_home, quite=True)
         # mv any existing conf file to backup
         conf = os.path.join(rmq_home, "etc/rabbitmq/rabbitmq.conf")
         if os.path.exists(conf):
             os.rename(conf, os.path.join(rmq_home,
-                                         "etc/rabbitmq/rabbitmq.conf_" +
+                                         "etc/rabbitmq/rabbitmq.conf_"+
                                          time.strftime("%Y%m%d-%H%M%S")
                                          ))
+    else:
+        filename = wget.download(
+            "https://github.com/rabbitmq/rabbitmq-server/releases/download/v3.7.7/rabbitmq-server-generic-unix-3.7.7.tar.xz",
+            out=os.path.expanduser("~"))
+        print("\nDownloaded rabbbitmq server")
+        print("\nExtracting {} to {}".format(filename, rmq_install_dir))
+        cmd = ["tar",
+               "-xf",
+               filename,
+               "--directory="+rmq_install_dir]
 
-    if is_valid_amqp_port(amqp_port) and is_valid_mgmt_port(mgmt_port):
-        # If ports are valid, add them to the rabbitmq conf
-        new_conf = """listeners.tcp.default = 5672
-            management.listener.port = 15672"""
+        subprocess.check_call(cmd)
+
+    config_opts['rmq-home'] = rmq_home
+
+
+
+    # Get vhost
+    vhost = config_opts.get('virtual-host', 'volttron')
+    prompt = 'What is the name of the virtual ' \
+             'host under which Rabbitmq VOLTTRON will be running?'
+    new_vhost = prompt_response(prompt, default=vhost)
+    config_opts['virtual-host'] = new_vhost
+    prompt = prompt_response('\nDo you want SSL Authentication',
+                             valid_answers=y_or_n,
+                             default='Y')
+    if prompt in y:
+        config_opts['ssl'] = "true"
+    else:
+        config_opts['ssl'] = "false"
+    config_opts['user'] = "guest"
+    config_opts['pass'] = "guest"
+    config_opts['host'] = 'localhost'
+
+    prompt = prompt_response('\nUse default rabbitmq ports ',
+                             valid_answers=y_or_n,
+                             default='Y')
+    if prompt in y:
+        config_opts['amqp-port'] = '5672'
+        config_opts['mgmt-port'] = '15672'
+        config_opts['rmq-address'] = build_rmq_address(ssl_auth=False,
+                                                       config=config_opts)
+        config_opts.sync()
+    else:
+        prompt = 'What is the instance port for the RabbitMQ address?'
+        prompt_port(config_opts, 'amqp-port', 5672, prompt)
+
+        prompt = 'What is the instance port for the RabbitMQ management plugin?'
+        prompt_port(config_opts, 'mgmt-port', 15672, prompt)
+        valid_port = False
+
+        config_opts.sync()
+        new_conf = """listeners.tcp.default = {}
+        management.listener.port = {}""".format(config_opts['amqp-port'],
+                                                config_opts['mgmt-port'])
         with open(os.path.join(config_opts['rmq-home'],
                                "etc/rabbitmq", "rabbitmq.conf"),
                   'w+') as r_conf:
             r_conf.write(new_conf)
 
-        # Start rabbitmq server
-        start_rabbit(config_opts['rmq-home'])
-    else:
-        return False
+
+    # Start rabbitmq server
+    start_rabbit(config_opts['rmq-home'])
+
+    #enable plugins
+    cmd =[os.path.join(config_opts['rmq-home'], "sbin/rabbitmq-plugins"),
+                       "enable", "rabbitmq_management",
+                       "rabbitmq_federation",
+                       "rabbitmq_federation_management",
+                       "rabbitmq_shovel", "rabbitmq_auth_mechanism_ssl"]
+    subprocess.check_call(cmd)
+    print("**enabled plugins")
 
 
-def _create_federation_setup():
+def prompt_port(config_opts, config_key, default_port, prompt):
+    # TODO - How to configure port other than 5671 for ssl - validate should
+    # check if port is not 5672.
+    port = config_opts.get(config_key, default_port)
+    valid_port = False
+    while not valid_port:
+        port = prompt_response(prompt, default=port)
+        if config_key == 'amqp-port':
+            valid_port = is_valid_amqp_port(port)
+        elif config_key == 'mgmt-port':
+            valid_port = is_valid_mgmt_port(port)
+        if not valid_port:
+            print("Port is not valid")
+    config_opts[config_key] = str(port)
+
+
+def _get_upstream_servers():
     """
-    Creates a RabbitMQ federation of multiple VOLTTRON instances based on rabbitmq config.
-        - Builds AMQP/S address for each upstream server
-        - Creates upstream servers
-        - Adds policy to make "volttron" exchange "federated".
-
+    Build AMQP/S URIs for upstream servers
     :return:
     """
-    global instance_name, config_opts
+    global config_opts, instance_name
+    if not config_opts:
+        _load_rmq_config()
+    federation = config_opts.get('federation-upstream', dict())
+    multi_platform = True
+    ssl_params = get_ssl_url_params()
+    prompt = 'How many upstream servers do you want to configure?'
+    count = prompt_response(prompt, default=1)
+    count = int(count)
+    i = 0
+    is_ssl = is_ssl_connection()
+    for i in range(0, count):
+        prompt = 'Name of the upstream server {}: '.format(i+1)
+        default_name = 'upstream-' + str(i+1)
+        name = prompt_response(prompt, default=default_name)
+        prompt = 'Hostname of the upstream server: '
+        host = prompt_response(prompt, default='localhost')
+        prompt = 'Port of the upstream server: '
+        port = prompt_response(prompt, default=5671)
+        prompt = 'Virtual host of the upstream server: '
+        vhost = prompt_response(prompt, default='volttron')
+        if is_ssl:
+            address = "amqps://{host}:{port}/{vhost}?" \
+                  "{ssl_params}&server_name_indication={host}".format(
+                    host=host, port=port, vhost=vhost, ssl_params=ssl_params)
+        else:
+            address = "amqp://{user}:{pwd}@{host}:{port}/{vhost}".format(
+                user=instance_name, pwd=instance_name, host=host, port=port,
+                vhost=vhost)
+        federation[name] = address
+    config_opts['federation-upstream'] = federation
+    config_opts.sync()
+    return multi_platform
+
+
+def _get_shovel_settings():
+    """
+    Prompt user for shovel information
+    :return:
+    """
+    global config_opts
     if not config_opts:
         _load_rmq_config()
 
-    federation = config_opts.get('federation-upstream')
-    if federation:
-        is_ssl = is_ssl_connection()
-        if is_ssl:
-            ssl_params = get_ssl_url_params()
-        try:
-            for upstream in federation:
-                print("Upstream Server: {host} ".format(upstream['host']))
-                name = "upstream-{host}".format(host=upstream['host'])
-                if is_ssl:
-                    address = "amqps://{host}:{port}/{vhost}?" \
-                              "{ssl_params}&server_name_indication={host}".format(
-                                host=upstream['host'],
-                                port=upstream['port'],
-                                vhost=upstream['virtual-host'],
-                                ssl_params=ssl_params)
-                else:
-                    address = "amqp://{user}:{pwd}@{host}:{port}/{vhost}".format(
-                                user=config_opts.get("user", instance_name+"-admin"),
-                                pwd=config_opts.get("pass", default_pass),
-                                host=upstream['host'],
-                                port=upstream['port'],
-                                vhost=upstream['virtual-host'])
-                prop = dict(vhost=config_opts['virtual-host'],
-                                component="federation-upstream",
-                                name=name,
-                                value={"uri":address})
-                set_parameter('federation-upstream',
-                              name,
-                              prop,
-                              config_opts['virtual-host'])
+    platform_config = load_platform_config()
+    try:
+        instance_name = platform_config['instance-name'].strip('"')
+        print(instance_name)
+    except KeyError as exc:
+        print("Unknown instance name. Please set instance-name in VOLTTRON_HOME/config")
+        return
 
-                policy_name = 'volttron-federation'
-                policy_value = {"pattern":"^volttron",
-                                "definition":{"federation-upstream-set":"all"},
-                                "priority":0,
-                                "apply-to": "exchanges"}
-                set_policy(policy_name, policy_value, config_opts['virtual-host'])
-        except KeyError as ex:
-            print("Federation setup  did not complete. "
-                  "Missing Key {key} in upstream config {upstream}".format(key=ex, upstream=upstream))
-            return ex
+    shovels = config_opts.get('shovels', [])
+    multi_platform = False
+    prompt = prompt_response('\nDo you want a multi-platform shovel setup? ',
+                                         valid_answers=y_or_n,
+                                         default='N')
+    if prompt in y:
+        multi_platform = True
+        prompt = prompt_response('\nDo you want shovels for multi-platform RPC? ',
+                                 valid_answers=y_or_n,
+                                 default='N')
+        if prompt in y:
+            prompt = 'How many RPC shovels do you want to configure? You will need to create one for each remote platform'
+            count = prompt_response(prompt, default=1)
+            for i in range(0, int(count)):
+                name = 'shovel-rpc-{}'.format(i + 1)
+                print("Configuring RPC Shovel {}".format(i+1))
+                prompt = 'Hostname of the destination instance: '
+                host = prompt_response(prompt, default='localhost')
+                prompt = 'Port of the upstream server: '
+                port = prompt_response(prompt, default=5672)
+                prompt = 'Virtual host of the destination server: '
+                vhost = prompt_response(prompt, default='volttron')
+                prompt = 'Username of the destination server: '
+                user = prompt_response(prompt, default='volttron')
+                prompt = 'Password of the destination server: '
+                pwd = prompt_response(prompt, default='volttron')
+                prompt = 'Instance name of the destination server: '
+                platform = prompt_response(prompt, default='')
+                address = "amqp://{0}:{1}@{2}:{3}/{4}".format(user, pwd, host, port, vhost)
+                rpc_key = platform + '.*'
+                shovels.append(dict(name=name, remote_address=address, topics=rpc_key))
+
+        prompt = prompt_response('\nDo you want shovels for multi-platform PUBSUB? ',
+                                 valid_answers=y_or_n,
+                                 default='N')
+        if prompt in y:
+            prompt = 'How many remote instances do you want to publish topic? '
+            count = prompt_response(prompt, default=1)
+            for i in range(0, int(count)):
+                print("Configuring Remote instance {}".format(i + 1))
+                prompt = 'Hostname of the destination instance: '
+                host = prompt_response(prompt, default='localhost')
+                prompt = 'Port of the upstream server: '
+                port = prompt_response(prompt, default=5672)
+                prompt = 'Virtual host of the destination server: '
+                vhost = prompt_response(prompt, default='volttron')
+                prompt = 'Username of the destination server: '
+                user = prompt_response(prompt, default='volttron')
+                prompt = 'Password of the destination server: '
+                pwd = prompt_response(prompt, default='volttron')
+                address = "amqp://{user}:{pwd}@{host}:{port}/{vhost}".format(
+                    user=user, pwd=pwd, host=host, port=port, vhost=vhost)
+                prompt = 'List of PUBSUB topics to publish to this remote instance (comma seperated)'
+                topics = prompt_response(prompt, default="")
+                topics = topics.split(",")
+
+                for topic in topics:
+                    name = "shovel-pubsub-{0}-{1}".format(topic, i+1)
+                    pubsub_key = "__pubsub__.{0}.{1}.#".format(instance_name, topic)
+                    shovels.append(dict(name=name, remote_address=address, topics=pubsub_key))
+        config_opts['shovel'] = shovels
+        config_opts.sync()
+
+    return multi_platform
 
 
 def _create_shovel_setup():
     """
-    Create RabbitMQ shovel based on the RabbitMQ config
+    Create RabbitMQ shovel based on the information provided by user
     :return:
     """
     global instance_name
     if not config_opts:
         _load_rmq_config()
-
+        return
+    platform_config = load_platform_config()
     shovels = config_opts.get('shovel', [])
-    is_ssl = is_ssl_connection()
-    src_uri = build_rmq_address(is_ssl)
-    if is_ssl:
-        ssl_params = get_ssl_url_params()
     print shovels
-    try:
-        for shovel in shovels:
-            # Build destination address
-            if is_ssl:
-                dest_uri = "amqps://{host}:{port}/{vhost}?" \
-                           "{ssl_params}&server_name_indication={host}".format(
-                           host=shovel['host'],
-                           port=shovel['port'],
-                           vhost=shovel['virtual-host'],
-                           ssl_params=ssl_params)
-            else:
-                dest_uri = "amqp://{user}:{pwd}@{host}:{port}/{vhost}".format(
-                    user=config_opts.get("user", instance_name + "-admin"),
-                    pwd=config_opts.get("pwd", default_pass),
-                    host=shovel['host'],
-                    port=shovel['port'],
-                    vhost=shovel['virtual-host'])
-
-            pubsub_topics = shovel.get("pubsub-topics", [])
-            agent_ids = shovel.get("rpc-agent-identities", [])
-            for topic in pubsub_topics:
-                print "Creating shovel to forward PUBSUB topic {}".format(topic)
-                name = "shovel-{host}-{topic}".format(host=shovel['host'],
-                                                      topic=topic)
-                routing_key = "__pubsub__.{instance}.{topic}.#".format(instance=instance_name,
-                                                                       topic=topic)
-                prop = dict(vhost=config_opts['virtual-host'],
-                                component="shovel",
-                                name=name,
-                                value={"src-uri": src_uri,
-                                        "src-exchange":  "volttron",
-                                        "src-exchange-key": routing_key,
-                                        "dest-uri": dest_uri,
-                                        "dest-exchange": "volttron"}
+    src_uri = build_rmq_address(ssl_auth=False)
+    for shovel in shovels:
+        name = shovel['name']
+        dest_uri = shovel['remote_address']
+        topics = shovel['topics']
+        if not isinstance(topics, list):
+            topics = [topics]
+        for topic in topics:
+            property = dict(vhost=config_opts['virtual-host'],
+                        component="shovel",
+                        name=name,
+                        value={"src-uri": src_uri,
+                                "src-exchange":  "volttron",
+                                "src-exchange-key": topic,
+                                "dest-uri": dest_uri,
+                                "dest-exchange": "volttron"}
                             )
-                print "shovel property: {}".format(prop)
-                # set_parameter("shovel",
-                #               name,
-                #               prop)
+            print("shovel property: {}", property)
+            set_parameter("shovel",
+                          name,
+                          property)
 
-            for identity in agent_ids:
-                print "Creating shovel to make RPC call to remote Agent : {}".format(topic)
-                name = "shovel-{host}-{identity}".format(host=shovel['host'],
-                                                         identity=identity)
-                routing_key = "{instance}.{identity}.#".format(instance=instance_name,
-                                                               identity=identity)
-                prop = dict(vhost=config_opts['virtual-host'],
-                                component="shovel",
-                                name=name,
-                                value={"src-uri": src_uri,
-                                        "src-exchange":  "volttron",
-                                        "src-exchange-key": routing_key,
-                                        "dest-uri": dest_uri,
-                                        "dest-exchange": "volttron"}
-                                )
-                print "shovel property: {}".format(prop)
-                # set_parameter("shovel",
-                #               name,
-                #               prop)
-    except KeyError as exc:
-        print("Shovel setup  did not complete. Missing Key: {}".format(exc))
+
+def wizard(type):
+    """
+    Setup VOLTTRON instance to run with RabbitMQ message bus.
+    :param type:
+            single - Setup to run as single instance
+            federation - Setup to connect multiple VOLTTRON instances as a federation
+            shovel - Setup shovels to forward local messages to remote instances
+    :return:
+    """
+    global instance_name
+    # First things first. Confirm VOLTTRON_HOME
+    print('\nYour VOLTTRON_HOME currently set to: {}'.format(get_home()))
+    prompt = '\nIs this the volttron instance you are attempting to ' \
+             'configure rabbitmq for? '
+    if not prompt_response(prompt, valid_answers=y_or_n, default='Y') in y:
+        print(
+            '\nPlease execute with VOLTRON_HOME=/your/path python '
+            'volttron/utils/rmq_mgmt.py to  '
+            'modify VOLTTRON_HOME.\n')
+        return
+
+    instance_name = get_platform_instance_name(prompt=True)
+    # Store config this is checked at startup
+    store_message_bus_config(message_bus='rmq', instance_name=instance_name)
+
+    if type == 'single':
+        # Get vhost from the user
+        _set_initial_rabbit_config(instance_name)
+        # Create local RabbitMQ setup
+        response = init_rabbitmq_setup() #should be called after config
+        # changes are written to disk.
+        ssl_auth = config_opts.get('ssl', "true")
+        if response and ssl_auth in ('true', 'True', 'TRUE'):
+            _setup_for_ssl_auth(instance_name)
+    elif type == 'federation':
+        _load_rmq_config()
+        # Create a federation setup
+        federation_needed = _get_upstream_servers()
+        if federation_needed:
+            _create_federation_setup()
+    elif type == 'shovel':
+        _load_rmq_config()
+        shovel_needed = _get_shovel_settings()
+        if shovel_needed:
+            _create_shovel_setup()
+    else:
+        print "Unknown option. Exiting...."
 
 
 def _setup_for_ssl_auth(instance_name):
     """
     Utility method to create
     1. Root CA
-    2. Instance CA
+    2. instance CA
     3. RabbitMQ server certificates (public and private)
-    4. Admin user to connect to RabbitMQ management web interface
+    4. RabbitMQ config with SSL setting
+    5. Admin user to connect to RabbitMQ management Web interface
 
     :param instance_name: Instance name
     :return:
@@ -310,8 +479,13 @@ def _setup_for_ssl_auth(instance_name):
     instance_ca_name, server_name, admin_client_name = \
         certs.Certs.get_cert_names(instance_name)
 
+    host = config_opts.get('host', 'localhost')
+    prompt = 'What is the fully qualified domain name of the system?'
+    new_host = prompt_response(prompt, default=getfqdn())
+    config_opts['host'] = new_host
+
     # prompt for host before creating certs as it is needed for server cert
-    _create_certs_without_prompt(admin_client_name, instance_ca_name, server_name)
+    _create_certs(admin_client_name, instance_ca_name, server_name)
 
     # if all was well, create the rabbitmq.conf file for user to copy
     # /etc/rabbitmq and update VOLTTRON_HOME/rabbitmq_config.json
@@ -339,18 +513,15 @@ management.listener.ssl_opts.keyfile = {server_key}""".format(
     with open(os.path.join(get_home(), "rabbitmq.conf"), 'w') as rconf:
         rconf.write(new_conf)
 
+    # updated rabbitmq_config.json
+    config_opts['user'] = admin_client_name
+    config_opts['pass'] = ""
+    config_opts['amqp-port'] = '5671'
+    config_opts['mgmt-port'] = '15671'
+    config_opts.sync()
+
     # Stop server, move new config file with ssl params, start server
     stop_rabbit(config_opts['rmq-home'])
-
-    # Change to SSL ports
-    config_opts["amqp-port"] = amqp_port = 5671 # TODO - If user defined port, this needs to change
-    config_opts["mgmt-port"] = mgmt_port = 15671 # TODO - If user defined port, this needs to change
-    # Add VOLTTRON admin user. This is needed to connect to web interface, multi-platform federation, shovel
-    config_opts["user"] = admin_client_name
-    config_opts["pass"] = default_pass
-
-    with open("{vhome}/rabbitmq_config.yml".format(vhome=get_home()), 'w') as yaml_file:
-        yaml.dump(config_opts, yaml_file, default_flow_style=False)
 
     os.rename(os.path.join(get_home(), "rabbitmq.conf"),
               os.path.join(config_opts.get("rmq-home"),
@@ -361,18 +532,11 @@ management.listener.ssl_opts.keyfile = {server_key}""".format(
           "{} and password={}.\nYou could change this user's password by "
           "logging into https://{}:{}/"
           "\n#######################".format(config_opts['user'],
-                                             default_pass,
-                                             config_opts['host'],
+                                             default_pass, config_opts['host'],
                                              config_opts['mgmt-port']))
 
 
 def stop_rabbit(rmq_home, quite=False):
-    """
-    Stop RabbitMQ Server
-    :param rmq_home: RabbitMQ installation path
-    :param quite:
-    :return:
-    """
     try:
         cmd = [os.path.join(rmq_home, "sbin/rabbitmqctl"),
                "stop"]
@@ -388,71 +552,11 @@ def stop_rabbit(rmq_home, quite=False):
 
 
 def start_rabbit(rmq_home):
-    """
-    Start RabbitMQ server
-    :param rmq_home: RabbitMQ installation path
-    :return:
-    """
     cmd = [os.path.join(rmq_home, "sbin/rabbitmq-server"),
            "-detached"]
     subprocess.check_call(cmd)
-    gevent.sleep(10)
+    gevent.sleep(4)
     print("**Started rmq server")
-
-
-def _create_certs_without_prompt(admin_client_name, instance_ca_name, server_cert_name):
-    """
-    Utility method to create certificates
-    :param client_cert_name: client (agent) cert name
-    :param instance_ca_name: VOLTTRON instance name
-    :param server_cert_name: RabbitMQ sever name
-    :return:
-    """
-    print admin_client_name
-    global config_opts
-    create_instance_ca = False
-    instance_ca_path = config_opts.get("instance_ca_path", None)
-    instance_ca_key = config_opts.get("instance_ca_key", None)
-    # Check if insance ca public and private certs exist and valid
-    found = _verify_and_save_instance_ca(instance_ca_path, instance_ca_key)
-    if found:
-        # get instance cert name
-        filename = os.path.basename(instance_ca_path)
-        instance_ca_name = os.path.splitext(filename)[0]
-    else:
-        # create ca cert in default dir if needed
-        if crts.ca_exists():
-            print('\n Root CA Found {}'.format(crts.cert_file(ROOT_CA_NAME)))
-            create_instance_ca = True
-        else:
-            print('\n Root CA NOT Found {}'.format(crts.cert_file(ROOT_CA_NAME)))
-            create_ca(override=False)
-            create_instance_ca = True
-
-        if crts.cert_exists(instance_ca_name):
-            create_instance_ca = False
-            instance_ca_path = crts.cert_file(instance_ca_name)
-            instance_ca_key = crts.private_key_file(instance_ca_name)
-        # create instance CA (intermediate CA) if root CA exists and instance cert
-        # doesn't exist
-        if create_instance_ca:
-            crts.create_instance_ca(instance_ca_name)
-        else:
-            found = _verify_and_save_instance_ca(instance_ca_path, instance_ca_key)
-            if found:
-                # get instance cert name
-                filename = os.path.basename(instance_ca_path)
-                instance_ca_name = os.path.splitext(filename)[0]
-
-    crts.create_ca_signed_cert(server_cert_name, type='server',
-                               ca_name=instance_ca_name,
-                               fqdn=config_opts.get('host'))
-
-    crts.create_ca_signed_cert(admin_client_name, type='client',
-                               ca_name=instance_ca_name)
-    create_user(admin_client_name, ssl_auth=False)
-    permissions = dict(configure=".*", read=".*", write=".*")
-    set_user_permissions(permissions, admin_client_name, ssl_auth=False)
 
 
 def _create_certs(client_cert_name, instance_ca_name, server_cert_name):
@@ -545,54 +649,15 @@ def _verify_and_save_instance_ca(instance_ca_path, instance_ca_key):
     return found
 
 
-def setup_rabbitmq_volttron(type):
-    """
-    Setup VOLTTRON instance to run with RabbitMQ message bus.
-    :param type:
-            single - Setup to run as single instance
-            federation - Setup to connect multiple VOLTTRON instances as a federation
-            shovel - Setup shovels to forward local messages to remote instances
-    :return:
-    """
-    global instance_name
-
-    instance_name = get_platform_instance_name(prompt=True)
-    # Store config this is checked at startup
-    store_message_bus_config(message_bus='rmq', instance_name=instance_name)
-
-    if type in ["all", "single"]:
-        error = _check_basic_rabbit_config()
-        if not error:
-            # Create local RabbitMQ setup
-            success = init_rabbitmq_setup() # should be called after check basic config
-            # changes are written to disk.
-            ssl_auth = config_opts.get('ssl', "true")
-            if success and ssl_auth in ('true', 'True', 'TRUE'):
-                _setup_for_ssl_auth(instance_name)
-            if type == "all":
-                # Create a multi-platform federation setup
-                error = _create_federation_setup()
-                if not error:
-                    # Create shovel setup
-                    _create_shovel_setup()
-    elif type == "federation":
-        # Create a multi-platform federation setup
-        _create_federation_setup()
-    elif type == "shovel":
-        # Create shovel setup
-        _create_shovel_setup()
-    else:
-        print "Unknown option. Exiting...."
-
-
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(
         formatter_class=argparse.RawTextHelpFormatter)
     parser.add_argument('type',
-                        help='Instance type: all, single, federation or shovel')
+                        help='Instance type: single, federation or shovel')
     args = parser.parse_args()
     type = args.type
     try:
-        setup_rabbitmq_volttron(type)
+        wizard(type)
     except KeyboardInterrupt:
         print "Exiting setup process"
+
